moved online fighter to separate class
